+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes to Calva.
 ## [Unreleased]
 
 - Fix: [Connecting to an external REPL does not work when local nrepl port exists](https://github.com/BetterThanTomorrow/calva/issues/2303)
+- [Avoid formatting when breaking up line comments](https://github.com/BetterThanTomorrow/calva/issues/2296)
 
 ## [2.0.397] - 2023-11-17